PROGRESS: navcontroller now stores generated navigation meshs.
MPActivityWalk now uses navigation to path around known navigation mesh.
MPActivityApproach now uses navigation to path towards it's destination

Added Tasks: Hotspots, Ghostwalk

+++ MPSpell.m
@@ -157,7 +157,7 @@
 		}
 	}
 	
-	PGLog(@"     == spellID[%@] id[%d] rank[%d]",name,  spellID, currentRank);
+	PGLog(@"     == spellID[%@] id[%d] rank[%d]",name,  (int)spellID, currentRank);
 	// end if
 }
